+++ src/js/domain/index.js
@@ -1,0 +1,4 @@
+export { default as Lotto } from './Lotto.js';
+export { default as Exchange } from './Exchange.js';
+export { default as LottoChecker } from './LottoChecker.js';
+export { default as LottoMachine } from './LottoMachine.js';
