+++ src/js/view/Lotto/index.js
@@ -1,0 +1,2 @@
+export { default as LottoInputView } from './LottoInputView.js';
+export { default as LottoOutputView } from './LottoOutputView.js';
